requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyserial==3.5
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
